Style labels for disabled inputs.
Summary:
The current styling for disabled inputs is very subtle.
It is clearer if the labels are also faded out.

Styling without requiring extra classes on the labels is tricky,
this is one attempt.

Please let me know what you think, and what the best way to do this
would be.

Reviewers: angel

Reviewed By: angel

Differential Revision: http://phabricator.bitnami.com:8080/D17850